Move conformance test assertions source to `org.jspecify.conformance.tests` package.
